One tokens clause for every harness's turn-complete note
The Codex note's cache split (previous commit) copied the shape OpenCode's note got
in #225, and two copies drift. `turn-usage-prose.ts` is the one statement, in the
fleet's `TurnTokenUsage` vocabulary at the harness root so it names no vendor; both
parsers read it and OpenCode re-exports `cacheDetail` for the tests that pinned it.
Byte-identical output on both feeds (pinned by their existing tests).
